Creating siepic_tools python package
- moving SiEPIC.utils.* into siepic_tools.utils.
