+++ klayout_dot_config/python/siepic_tools/__init__.py
@@ -1,0 +1,7 @@
+# Detecting environment
+
+try:
+    import pya
+    is_GSI = pya.__spec__.origin == 'built-in'
+except ImportError:
+    is_GSI = False
