+++ klayout_dot_config/python/siepic_tools/config.py
@@ -1,0 +1,1 @@
+PIN_LENGTH = 100  # 0.1 micron
